requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.0
cssselect2==0.7.0
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.0.0
django-filter==23.1
django-jazzmin==2.6.0
django-statsd-mozilla==0.4.0
djangorestframework==3.14.0
filelock==3.9.0
googlemaps==4.10.0
html5lib==1.1
idna==3.4
joblib==1.2.0
lxml==4.9.2
Markdown==3.4.3
mysqlclient==2.1.1
nltk==3.7
oscrypto==1.3.0
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.0
psycopg2==2.9.5
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.6.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
regex==2022.10.31
reportlab==3.6.12
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
statsd==4.0.0
svglib==1.5.1
tenacity==8.2.1
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.14
virtualenv==20.19.0
webencodings==0.5.1
xhtml2pdf==0.2.9
